(refactor) Se sacó el RegistryQuery de la carpeta de entities
