+++ src/contexts/QrAttendance/registry/domain/registry.query.ts
